+++ docs/examples/plugins/pydantic_fast.py
@@ -1,0 +1,19 @@
+import pydantic
+
+from dmr import Body, Controller
+from dmr.plugins.pydantic import PydanticFastSerializer
+
+
+class _User(pydantic.BaseModel):
+    username: str
+    age: int
+
+
+class UserController(Controller[PydanticFastSerializer]):
+    def put(self, parsed_body: Body[_User]) -> _User:
+        return parsed_body
+
+
+# run: {"controller": "UserController", "url": "/api/users/", "method": "put", "body": {"username": "sobolevn", "age": 27}}  # noqa: ERA001, E501
+# run: {"controller": "UserController", "url": "/api/users/", "method": "put", "body": {"username": "sobolevn"}, "curl_args": ["-D", "-"], "assert-error-text": "age", "fail-with-body": false}  # noqa: ERA001, E501
+# openapi: {"controller": "UserController", "openapi_url": "/docs/openapi.json/"}  # noqa: ERA001, E501