+++ tests/test_unit/test_plugins/test_pydantic/test_pydantic_snapshots.py
@@ -6,7 +6,7 @@ from syrupy.assertion import SnapshotAssertion
 
 from dmr import Body, Controller
 from dmr.openapi import build_schema
-from dmr.plugins.pydantic import PydanticSerializer
+from dmr.plugins.pydantic import PydanticFastSerializer, PydanticSerializer
 from dmr.routing import Router
 
 
@@ -20,6 +20,11 @@ class _UserController(Controller[PydanticSerializer]):
         raise NotImplementedError
 
 
+class _UserFastController(Controller[PydanticFastSerializer]):
+    def post(self, parsed_body: Body[_UserModel]) -> str:
+        raise NotImplementedError
+
+
 def test_user_schema(snapshot: SnapshotAssertion) -> None:
     """Ensure that schema is correct for user controller."""
     assert (
@@ -27,7 +32,10 @@ def test_user_schema(snapshot: SnapshotAssertion) -> None:
             build_schema(
                 Router(
                     'api/v1/',
-                    [path('/user', _UserController.as_view())],
+                    [
+                        path('/user', _UserController.as_view()),
+                        path('/user-fast', _UserFastController.as_view()),
+                    ],
                 ),
             ).convert(),
             indent=2,